engine/requirements.txt
anyio==4.11.0
argcomplete==3.6.3
astroid==4.0.2
attrs==25.4.0
beautifulsoup4==4.14.3
black==25.12.0
blivet==3.12.1
Brlapi==0.8.7
certifi==2025.7.9
cffi==2.0.0
charset-normalizer==3.4.3
click==8.1.7
cockpit @ file:///builddir/build/BUILD/cockpit-351-build/cockpit-351/tmp/wheel/cockpit-351-py3-none-any.whl#sha256=59d2dcd42c86f8d88df2a0b4afd6bee5c58753efde42f78d6ed1686b14aa61df
crypt_r==3.13.1
cupshelpers==1.0
dasbus==1.7
dbus-python==1.4.0
dbus_next==0.2.3
dill==0.4.0
distro==1.9.0
dnf==4.24.0
fedora-third-party==0.10
file-magic==0.4.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
humanize==4.13.0
idna==3.10
iso639==0.1.4
isort==7.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langtable==0.0.69
libcomps==0.1.23
libdnf==0.75.0
libdnf5==5.2.17.0
louis==3.33.0
lxml==6.0.1
mccabe==0.7.0
mercurial==7.1.1
mypy_extensions==1.1.0
nftables==0.1
olefile==0.47
packaging==25.0
pathspec==0.12.1
perf==0.1
pexpect==4.9.0
pid==3.0.4
pillow==11.3.0
platformdirs==4.5.1
ply==3.11
productmd==1.49
progressbar2==4.5.0
ptyprocess==0.7.0
pwquality==1.4.5
PyAudio==0.2.13
pycairo==1.28.0
pycparser==2.22
pycups==2.0.4
pyenchant==3.2.2
PyGObject==3.54.5
pyinotify==0.9.6
pykickstart @ file:///builddir/build/BUILD/pykickstart-3.66-build/pykickstart-3.66
pylint==4.0.4
pyparted==3.13.0
PySocks==1.7.1
python-augeas==1.2.0
python-dateutil==2.9.0.post0
python-linux-procfs==0.7.3
python-meh==0.52
python-nmap==0.7.1
python-pam==2.0.2
python-utils==3.9.1
pytokens==0.3.0
pyudev==0.24.3
pyxdg==0.27
PyYAML==6.0.2
pyynl @ file:///builddir/build/BUILD/kernel-6.17.11-build/kernel-6.17.11/linux-6.17.11-300.fc43.x86_64/tools/net/ynl
RapidFuzz==3.12.2
referencing==0.36.2
regex==2025.10.23
requests==2.32.5
requests-file==2.0.0
requests-ftp==0.3.1
rpds-py==0.27.0
rpm==6.0.0
rpmautospec==0.8.3
rpmautospec-core==0.1.5
selinux @ file:///builddir/build/BUILD/libselinux-3.9-build/libselinux-3.9/src
sepolicy @ file:///builddir/build/BUILD/policycoreutils-3.9-build/selinux-3.9/python/sepolicy
setools==4.6.0
simpleline==1.9.0
six==1.17.0
sniffio==1.3.1
sos==4.10.1
soupsieve==2.8
systemd-python==235
tomlkit==0.13.3
typing_extensions==4.15.0
ufw==0.35
urllib3==2.6.2
xkbregistry==0.3
